restricy unauthorized or unauthenticated users from the view level

+++ base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.http import HttpResponse
 from django.contrib.auth.models import User
@@ -50,6 +51,7 @@ def room(request, id):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login_view')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -60,9 +62,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login_view')
 def updateRoom(request, id):
     room = Room.objects.get(id=id)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room.')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -71,8 +78,13 @@ def updateRoom(request, id):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login_view')
 def deleteRoom(request, id):
     room = Room.objects.get(id=id)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home_view')
